# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "md-view-my-collection API",
    "version": "0.1.0",
    "description": "Markdown レポートの保存・閲覧 UI の BFF（Next.js Route Handlers）。このドキュメントは zod スキーマ（front/src/schemas）から自動生成される。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同一オリジン（Next.js）"
    }
  ],
...
  "paths": {
...
    "/api/auth/admin": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "管理者判定",
        "description": "Bearer トークンを検証し ADMIN_EMAIL と照合する。",
        "responses": {
          "200": {
            "description": "判定結果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsAdminResponse"
                }
              }
            }
          },
          "401": {
            "description": "未認証 / トークン無効",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsAdminResponse"
                }
              }
            }
          },
          "429": {
            "description": "レートリミット超過",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/is-allowed": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "許可メール判定",
        "description": "local モードは body.email、supabase モードは Bearer トークンを検証する。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IsAllowedRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "判定結果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsAllowedResponse"
                }
              }
            }
          },
          "401": {
            "description": "未認証 / トークン無効",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsAllowedResponse"
                }
              }
            }
          },
          "429": {
            "description": "レートリミット超過",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "サーバー設定不備",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsAllowedResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "IsAllowedRequest": {
        "type": "object",
        "properties": {
          "email": {
            "description": "local モードで照合するメール",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "example": {
          "error": "Not found"
        }
      },
...
      "IsAdminResponse": {
        "type": "object",
        "properties": {
          "isAdmin": {
            "type": "boolean"
          }
        },
        "required": [
          "isAdmin"
        ],
        "additionalProperties": false
      },
      "IsAllowedResponse": {
        "type": "object",
        "properties": {
          "allowed": {
            "type": "boolean"
          }
        },
        "required": [
          "allowed"
        ],
        "additionalProperties": false
      }
    }
  }
}